Implement simple piped communication for Windows
This adds a simple version of the piped communication code for
Windows. This includes the creation of a new child process using
the constructor argument, and then redirecting STDIN/STDOUT/STDERR
so the parent can read from these.

This current works for some basic examples (unit tests to come), but
is mostly a proof of concept. Various aspects may need to be modified.
For this reason some code related to named pipes has been left in (but
commented out) as this is an alternative path that we may consider.