"Fix" unit tests for make test count checks
This unifies the number of unit tests on all OSs by putting ifdefs
inside test cases. Some unit tests are not done properly in this commit.
This is WIP to improve and check the CI results, not a final example of
the tests (i.e. commented out broken parts are known to be wrong and
many unit tests do not really test correctly on some OSs).