Pass correct output stream and fix usage for object_size
This changes the define_object_size function to take an
ostream argument since different invocations of the
SMT code can specify different ostreams. This led to
inconsistency where the option to print to a file
had assertions required for object_size calculations,
but these assertions did NOT appear in the SMT sent
to (e.g.) z3 when done via a temporary file.

Note that this is only needed here, but some local
conversion has been added because convert_expr(...) uses
the wrong ostream, and so symbols needed to be emitted
locally for the object_size.