Write object sizes once to the common out stream
Object sizes should be written after all other SMT output
but before the footer (which handles the actula sat query).
For sanity reasons the object sizes should be written to the
common outstream (this means they appear the same in all
calls to the SMT solver). This code change ensures they are
written once.

Note that there are two other ways this has been considered:
1. was to write them in the footer, but this requires passing
   the footer stream around.
2. was to write the footer to the common ostream, but this
   causes an infinite loop when multiple calls use the same
   base SMT and different st/unsat calls.