Ignore quantifier errors running `get-value` commands with `z3`
Due to how we call z3 it is possible to have values that contain
quantifiers appear in the results. However, z3 cannot give a
real value in response to a "(get-value |XXX|)" request.
This catches these any simply ignores them, allowing the rest
of the code to continue.

Note that this becomes necessary to handle since cbmc now
produces quantifiers in expressions sent to z3.